Init
Implement multiple platforms (with a slightly smaller slope above the main one).
After removing the cube, a bomb is created in its place, which smoothly becomes transparent (each bomb has a time of 2 to 5 seconds) and then explodes (scattering objects within a specified radius).

Each spawner has an indicator for displaying the following parameters on the screen
1) the number of spawned objects for the entire time (appearing on the scene)
2) the number of objects created
3) the number of active objects on the scene (Data can be taken from the Object Pool)
Information should be shown separately for cubes and separately for bombs